Move cross-app testing
