+++ .gitignore
@@ -7,6 +7,7 @@ wheels/
 *.egg-info
 
 .venv
+.vscode
 
 *.sqlite3*
