--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-    "editor.formatOnSave": true,
-    "editor.formatOnPaste": true,
-    "editor.detectIndentation": false,
-}
